+++ versions.tf
@@ -1,6 +1,6 @@
 terraform {
   experiments      = [module_variable_optional_attrs]
-  required_version = ">= 0.14"
+  required_version = ">= 0.14, < 1.3"
 
   required_providers {
     azurerm = {